requirements.txt
faiss-cpu
transformers
sentence-transformers
PyMuPDF
torch
scikit-learn
streamlit
